Extract code in separate classes.
