+++ TelemetryExporter.Console/Widgets/Interfaces/IWidget.cs
@@ -1,0 +1,9 @@
+﻿using TelemetryExporter.Console.Models;
+
+namespace TelemetryExporter.Console.Widgets.Interfaces
+{
+    internal interface IWidget
+    {
+        void GenerateImage(SessionData sessionData, FrameData currentData);
+    }
+}
